Rename from "renderer support" to "project reader"
