+++ ldj-project-reader/src/main/java/com/infosupport/ldoc/reader/Class.java
@@ -1,0 +1,11 @@
+package com.infosupport.ldoc.reader;
+
+import java.util.stream.Stream;
+
+public interface Class extends Type {
+
+  String fullName();
+
+  Stream<Method> methods();
+
+}
