+++ ldj-project-reader/src/main/java/com/infosupport/ldoc/reader/DocumentationComment.java
@@ -1,0 +1,6 @@
+package com.infosupport.ldoc.reader;
+
+public interface DocumentationComment {
+
+  String summary();
+}
