+++ ldj-project-reader/src/main/java/com/infosupport/ldoc/reader/Enum.java
@@ -1,0 +1,5 @@
+package com.infosupport.ldoc.reader;
+
+public interface Enum extends Type {
+
+}
