+++ ldj-project-reader/src/main/java/com/infosupport/ldoc/reader/Field.java
@@ -1,0 +1,5 @@
+package com.infosupport.ldoc.reader;
+
+public interface Field extends Node {
+
+}
