+++ ldj-project-reader/src/main/java/com/infosupport/ldoc/reader/Method.java
@@ -1,0 +1,9 @@
+package com.infosupport.ldoc.reader;
+
+public interface Method extends Node {
+
+  String name();
+
+  DocumentationComment documentationComment();
+
+}
