+++ ldj-project-reader/src/main/java/com/infosupport/ldoc/reader/Node.java
@@ -1,0 +1,6 @@
+package com.infosupport.ldoc.reader;
+
+public interface Node {
+  void accept(Visitor v);
+
+}
