+++ ldj-project-reader/src/main/java/com/infosupport/ldoc/reader/ProjectFactory.java
@@ -1,0 +1,17 @@
+package com.infosupport.ldoc.reader;
+
+import java.io.File;
+import java.io.IOException;
+import java.io.Reader;
+import java.net.URL;
+
+public interface ProjectFactory {
+
+  Project project(File file) throws IOException;
+
+  Project project(String string) throws IOException;
+
+  Project project(Reader reader) throws IOException;
+
+  Project project(URL url) throws IOException;
+}
