+++ ldj-project-reader/src/main/java/com/infosupport/ldoc/reader/Visitor.java
@@ -1,0 +1,12 @@
+package com.infosupport.ldoc.reader;
+
+public interface Visitor {
+
+  void visitClass(Class klass);
+
+  void visitMethod(Method method);
+
+  void visitProject(Project project);
+
+  void visitType(Type type);
+}
