+++ ldj-project-reader/src/main/java/com/infosupport/ldoc/reader/impl/UnknownTypeImpl.java
@@ -1,0 +1,12 @@
+package com.infosupport.ldoc.reader.impl;
+
+import com.infosupport.ldoc.reader.Type;
+import com.infosupport.ldoc.reader.Visitor;
+
+class UnknownTypeImpl implements Type {
+
+  @Override
+  public void accept(Visitor v) {
+    /* Do nothing. */
+  }
+}
