+++ ldj-project-reader/src/test/resources/com/infosupport/ldoc/reader/example.json
