--- ldj-renderer-support/src/main/java/com/infosupport/ldoc/support/Class.java
@@ -1,11 +1,0 @@
-package com.infosupport.ldoc.support;
-
-import java.util.stream.Stream;
-
-public interface Class extends Type {
-
-  String fullName();
-
-  Stream<Method> methods();
-
-}
