--- ldj-renderer-support/src/main/java/com/infosupport/ldoc/support/DocumentationComment.java
@@ -1,6 +1,0 @@
-package com.infosupport.ldoc.support;
-
-public interface DocumentationComment {
-
-  String summary();
-}
