--- ldj-renderer-support/src/main/java/com/infosupport/ldoc/support/Enum.java
@@ -1,5 +1,0 @@
-package com.infosupport.ldoc.support;
-
-public interface Enum extends Type {
-
-}
