--- ldj-renderer-support/src/main/java/com/infosupport/ldoc/support/Field.java
@@ -1,5 +1,0 @@
-package com.infosupport.ldoc.support;
-
-public interface Field extends Node {
-
-}
