--- ldj-renderer-support/src/main/java/com/infosupport/ldoc/support/Method.java
@@ -1,9 +1,0 @@
-package com.infosupport.ldoc.support;
-
-public interface Method extends Node {
-
-  String name();
-
-  DocumentationComment documentationComment();
-
-}
