--- ldj-renderer-support/src/main/java/com/infosupport/ldoc/support/Node.java
@@ -1,6 +1,0 @@
-package com.infosupport.ldoc.support;
-
-public interface Node {
-  void accept(Visitor v);
-
-}
