--- ldj-renderer-support/src/main/java/com/infosupport/ldoc/support/Project.java
@@ -1,12 +1,0 @@
-package com.infosupport.ldoc.support;
-
-import java.util.stream.Stream;
-
-public interface Project extends Node {
-  Stream<Type> allTypes();
-
-  Stream<Class> classes();
-  Stream<Interface> interfaces();
-  Stream<Enum> enums();
-  Stream<Struct> structs();
-}
