--- ldj-renderer-support/src/main/java/com/infosupport/ldoc/support/Struct.java
@@ -1,5 +1,0 @@
-package com.infosupport.ldoc.support;
-
-public interface Struct extends Type {
-
-}
