--- ldj-renderer-support/src/main/java/com/infosupport/ldoc/support/Visitor.java
@@ -1,12 +1,0 @@
-package com.infosupport.ldoc.support;
-
-public interface Visitor {
-
-  void visitClass(Class klass);
-
-  void visitMethod(Method method);
-
-  void visitProject(Project project);
-
-  void visitType(Type type);
-}
