+++ pom.xml
@@ -13,8 +13,8 @@
     <module>jacoco-report-aggregator</module>
     <module>ldj-maven-plugin</module>
     <module>ldj-maven-plugin-example</module>
-    <module>ldj-renderer-support</module>
     <module>ldj-gradle-plugin</module>
+    <module>ldj-project-reader</module>
     <module>ldj-spring-amqp-example</module>
     <module>ldj-spring-event-example</module>
     <module>ldj-spring-renderer-example</module>
